ruff_formatter: remove 'unsafe' used for non-UB concerns
In Rust, 'unsafe' has a very specific meaning: it delineates the
possibility of undefined behavior. It is not meant to be used as a lint
to discourage folks from using something. In this case, it looks like it
was used to make sure folks were careful when calling some APIs that
could panic but not lead to undefined behavior.

This commit removes the use of 'unsafe' for this purpose and buttons up
the docs a little bit.