ruff_python_literal: remove unchecked string conversion
It's unclear how perf is impacted by this change, but this particular
use of unsafe is unsound. Namely, `unsafe` is used inside of a safe
routine but there is no actual guarantee that the bytes being converted
to a string are valid UTF-8. Namely, anyone can do something like this:

    AsciiEscape::new(&[0xFF], EscapeLayout { .. }).write_source(..)

and get UB without uttering `unsafe`.

It looks like this was introduced when the code was initially written:
https://github.com/astral-sh/RustPython-Parser/commit/80109b1fe05ddf7fe9594d5132632c5e7430ca3b